\name{loadLibraryAs}
\alias{loadLibraryAs}
\title{Loads a R object with a given name into a desired environment}
\usage{
loadLibraryAs(name, filename, envir = globalenv())
}
\arguments{
  \item{name}{Name of the object you want to load}

  \item{filename}{Path to the file and name of the file
  that should be loaded}

  \item{enir}{The environment in which the object should be
  loaded}
}
\value{
Tthe name of the just created object
}
\description{
Loads a R object with a given name into a desired
environment
}

