\name{orderTwoColorArray}
\alias{orderTwoColorArray}
\title{Order by a given name}
\usage{
orderTwoColorArray(x, by = "Name")
}
\arguments{
  \item{x}{Object that is two dimensional (x[ordering,] is
  called)}

  \item{by}{The variable to order from}
}
\value{
An ordered object
}
\description{
Order by a given name
}

