\name{plotBoxDensityFromRG}
\alias{plotBoxDensityFromRG}
\title{A plot is created as a combination of Boxplot and Density plot of the array intensities}
\usage{
plotBoxDensityFromRG(RG)
}
\arguments{
  \item{RG}{Object of type RGList}
}
\value{
Already performs plotting
}
\description{
A plot is created as a combination of Boxplot and Density
plot of the array intensities
}

